feat: added permissions
