+++ src/components/ActionsBuilder/AddEditOptionModal/AddEditOptionModal.tsx
@@ -41,6 +41,7 @@ export const AddEditOptionModal: React.FC<AddEditOptionModalProps> = ({
         label,
         color: theme?.colors?.votes?.[options.length + 1],
         decodedActions: [],
+        permissions: [],
       },
     ];
     onChange(newOptions);
