+++ src/hooks/Guilds/contracts/useEncodedCall.ts
@@ -1,4 +1,5 @@
 import { Call, DecodedCall, Option } from 'components/ActionsBuilder/types';
+import { permissionChecks } from 'components/ActionsBuilder/utils';
 import ERC20 from 'contracts/ERC20.json';
 import { utils, BigNumber } from 'ethers';
 
@@ -53,6 +54,7 @@ export const bulkEncodeCallsFromOptions = (options: Option[]): Option[] => {
     return {
       ...option,
       actions: encodedCalls,
+      permissions: permissionChecks(decodedActions),
     };
   });
 };
